W175b: btlTarget empties its over-window set -- FOUR clears, ~400 bytes of invented work deleted, one function left at nd2.
  002D7560   964/816  nd766  ->  816/816  nd128   duplicate HAND-WRITTEN unsigned-float conversion removed, reward entries/fields typed, sparse switch layout restored
  002D8390   656/640  nd343  ->  628/640  **nd2**  semantic roles separated, ASCENDING state-case layout restored, typed enemy-table entry address staged
  002D8F70  2992/2864 nd2134 -> 2816/2864 nd2004  INVENTED post-call truncation deleted -- it was forcing repeated zero-extension through every case
  002D9D70  2724/2720 nd1986 -> 2720/2720 nd1974  pointer-sized helper result recovered as u32 instead of undefined8
  002D8110   508/512  nd339  ->  496/512  nd286   u32 helper ABI, two-value dispatch rebuilt as a typed switch

002D7560 is a direct vindication of the call I made an hour ago on bcm_panel 0022C2D0, where an agent's best result required HAND-WRITING the unsigned float-to-integer expansion and I refused it as a source-honesty violation. Here the same construct was found already present in the source and DELETING it moved the function 638 points and cleared 148 bytes. **Hand-written unsigned-float conversion is invented work, full stop -- the native cast is what retail's compiler emitted.**

002D8F70 names a defect class not seen before: an **invented post-call TRUNCATION** whose cost was not local. It forced repeated zero-extension at every subsequent use across every switch case, so one wrong narrowing cost 176 bytes spread across a 2.9 KB function. Worth checking wherever a call result is narrowed immediately on return.

002D8390 at nd2 is the file's best remaining target; its entire residual is two register-allocation words where slti/bnez use $v0 instead of retail's $at.

123/146, over-window 4 -> 0, no lost MATCHes, CRLF-only preserved, zero optimization pragmas in the file.

+++ src/Event/Fcl/fcl_misc.c
@@ -12,7 +12,7 @@ extern u32 FUN_003c9460(int *);
 #pragma alias fclMisc9240Call FUN_003c9240
 extern void fclMisc9240Call(int *);
 #pragma alias fclMisc9290Call FUN_003c9290
-extern u64 fclMisc9290Call(int *,u64,u64);
+extern u64 fclMisc9290Call(int *,u32,u32);
 #pragma alias fclMisc9340Call FUN_003c9340
 extern u32 fclMisc9340Call(int *);
 #pragma alias fclMisc9390Call FUN_003c9390
